+++ main/exploration.h
@@ -89,6 +89,6 @@ OUTPUT: Update maze for learned walls
 */
 void floodFillUpdate(coord currCoord, int heading, int *lfr);
 
-void exploration(Queue& motion_queue, int *lfr, int *startStep, String keyword, int *cellCount, int *heading);
+void exploration(Queue& motion_queue, int *lfr, int *startStep, String keyword, int *cellCount, byte *heading);
 
 #endif
